ms-agent-harness/requirements.txt
# MS Agent Harness — Dependencies
# Install: pip install -r requirements.txt

# Microsoft Agent Framework (consolidated AutoGen + Semantic Kernel)
agent-framework>=0.1.0
agent-framework-foundry>=0.1.0       # Azure AI Foundry chat client
agent-framework-openai>=0.1.0        # Direct OpenAI fallback for local dev
agent-framework-orchestrations>=0.1.0 # SequentialBuilder, HandoffBuilder

# Azure
azure-identity>=1.16.0              # DefaultAzureCredential (Managed Identity / CLI)
azure-storage-blob>=12.19.0         # State persistence to Blob Storage

# REST API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0

# HTTP client (ADO REST API calls)
httpx>=0.27.0

# Persistence
aiosqlite>=0.20.0                   # SQLite async for analysis cache + checkpoint/resume

# Configuration
pyyaml>=6.0.0                       # settings.yaml parsing

# ── Test dependencies ──
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.24.0
